clean up repair bag code, add untracked file deletion feature

+++ ami_bag/update_bag.py
@@ -4,18 +4,19 @@ import os
 import shutil
 import datetime
 import re
+import logging
 
 SYSTEM_FILE_PATTERNS = {
     "Thumbs.db": {
-        "regex": r"[Tt]humbs\\.db$",
+        "regex": r"[Tt]humbs\.db$",
         "match": False
     },
     "DS_Store": {
-        "regex": r"\\.DS_Store$",
+        "regex": r"\.DS_Store$",
         "match": False
     },
     "Appledouble": {
-        "regex": r"\\._.+$",
+        "regex": r"\._.+$",
         "match": False
     },
     "Icon": {
@@ -24,12 +25,14 @@ SYSTEM_FILE_PATTERNS = {
     }
 }
 
+LOGGER = logging.getLogger(__name__)
 
 class Repairable_Bag(bagit.Bag):
 
   def __init__(self, *args, **kwargs):
     super(Repairable_Bag, self).__init__(*args, **kwargs)
     self.old_dir = os.path.abspath(os.path.curdir)
+
 
   def check_baginfo(self):
     try:
@@ -66,7 +69,7 @@ class Repairable_Bag(bagit.Bag):
     try:
       bagit._make_tag_file(os.path.join(self.path, "bag-info.txt"), self.info)
     except:
-      print("Do not have permission to overwrite bag-info")
+      LOGGER.error("Do not have permission to overwrite bag-info")
       return False
 
     for alg in set(self.algs):
@@ -76,9 +79,8 @@ class Repairable_Bag(bagit.Bag):
 
   '''
   def remove_manifestentry(self):
-
-  def remove_payload(self):
   '''
+
 
   def payload_files_not_in_manifest(self):
     """
@@ -88,6 +90,7 @@ class Repairable_Bag(bagit.Bag):
       for payload_file in self.compare_manifests_with_fs()[1]:
          yield payload_file
 
+
   def add_payload_files_not_in_manifest(self):
     """
     iterate through all dem new files
@@ -95,13 +98,18 @@ class Repairable_Bag(bagit.Bag):
     os.chdir(self.path)
 
     new_payload_files = list(self.payload_files_not_in_manifest())
-    for payload_file in self.payload_files_not_in_manifest():
-      #print payload_file
-      self.add_payload_file_to_manifest(payload_file)
 
-    self.update_baginfo()
-    self.copy_manifest_files()
-    self.rewrite_manifest_files()
+    if new_payload_files:
+      LOGGER.info("Adding the following files to manifests: {}".format(", ".join(new_payload_files)))
+      for payload_file in self.payload_files_not_in_manifest():
+        self.add_payload_file_to_manifest(payload_file)
+
+      try:
+        self.copy_manifest_files()
+        self.rewrite_manifest_files()
+        self.update_baginfo()
+      except:
+        LOGGER.error("Do not have permission to write new manifests")
 
     os.chdir(self.old_dir)
 
@@ -118,7 +126,7 @@ class Repairable_Bag(bagit.Bag):
     self.entries[filename] = stuff
 
 
-  def delete_payload_files_not_in_manifest(self, rules = None):
+  def delete_payload_files_not_in_manifest(self, rules = SYSTEM_FILE_PATTERNS):
     """
     iterate through all dem new files
     """
@@ -126,29 +134,33 @@ class Repairable_Bag(bagit.Bag):
 
     if rules:
       files_to_delete = list()
+      files_not_to_delete = list()
       for payload_file in new_payload_files:
         for rulename, rule in rules.items():
           if bool(re.search(rule["regex"], payload_file)) != rule["match"]:
             files_to_delete.append(payload_file)
+        if payload_file not in files_to_delete:
+          files_not_to_delete.append(payload_file)
     else:
       files_to_delete = new_payload_files
 
+    if files_not_to_delete:
+      LOGGER.warning("Untracked files in payload directory do not match deletion rules: {}".format(", ".join(files_not_to_delete)))
+
     os.chdir(self.path)
+    if files_to_delete:
+      LOGGER.warning("Will delete the following files: {}".format(", ".join(files_to_delete)))
+      for payload_file in files_to_delete:
+        try:
+          LOGGER.warning("Deleting {}".format(payload_file))
+          os.remove(payload_file)
+        except OSError:
+          LOGGER.error("Do not have permission to delete {}".format(payload_file))
 
-    for payload_file in files_to_delete:
-      #print payload_file
-      try:
-        os.remove(payload_file)
-        #logging.info("Deleted {0}".format(path))
-      except OSError:
-
-        "eh"
-        #logging.error("Could not delete {0}".format(path))
+      if not self.check_baginfo():
+        self.update_baginfo()
 
     os.chdir(self.old_dir)
-
-    self.update_baginfo()
-    self.rewrite_manifest_files()
 
 
   def copy_manifest_files(self):
@@ -160,10 +172,16 @@ class Repairable_Bag(bagit.Bag):
 
   def rewrite_manifest_files(self):
     for alg in set(self.algs):
-      with open('manifest-%s.txt' % alg, 'w') as manifest:
-        for payload_file, hashes in self.entries.items():
-          if payload_file.startswith("data" + os.sep):
-            manifest.write("%s  %s\n" % (hashes[alg], bagit._encode_filename(payload_file)))
+      try:
+        with open('manifest-%s.txt' % alg, 'w') as manifest:
+          for payload_file, hashes in self.entries.items():
+            if payload_file.startswith("data" + os.sep):
+              manifest.write("%s  %s\n" % (hashes[alg], bagit._encode_filename(payload_file)))
+      except:
+        LOGGER.error("Do not have permission to overwrite hash manifests")
 
     for alg in set(self.algs):
-      bagit._make_tagmanifest_file(alg, self.path)
+      try:
+        bagit._make_tagmanifest_file(alg, self.path)
+      except:
+        LOGGER.error("Do not have permission to overwrite tag manifests")
